// src/enums.ts
export enum Environment {
    Dev = 'development',
    Prod = 'production'
};

export enum EFeatureType {
    ACTIONS_MAIN = 'sale_actions_category',
    ACTIONS_ADDITIONAL = 'sale_actions_stream',
    NEW_DESKTOP = 'new',
    NEW_MOBILE = 'new_mobile',
    HIT_DESKTOP = 'hit',
    HIT_MOBILE = 'hit_mobile'
};

export enum EModalType {
    NEW_HIT = 'newHit',
    ACTIONS_MAIN = 'actionsMain',
    ACTIONS_ADDITIONAL = 'actionsAdditional'
};